LayoutTests/http/tests/dom/window-open-about-webkit-org-and-access-document.html

<!DOCTYPE html>
<html lang="en">
<head>
    <title>Tests opening a new about://webkit.org window and accessing its document</title>
    <script src="../resources/js-test-pre.js"></script>
    <script>
        var newWindow;

        if (window.testRunner) {
            testRunner.setCanOpenWindows();
            testRunner.waitUntilDone();
        }

        function checkNewWindowDocumentIsUndefined () {
            shouldBeUndefined("newWindow.document");
            if (window.testRunner)
                testRunner.notifyDone();
        }

        function clickHandler() {
            newWindow = window.open("about://webkit.org");
            try {
                newWindow.document.write("<scri" + "pt>console.log('Injected script running.')</sc" + "ript>");
                testFailed("Was able to write to the new window's document.");
                if (window.testRunner)
                    testRunner.notifyDone();
            } catch (e) {
                setTimeout(checkNewWindowDocumentIsUndefined, 500);
            }
        }

        function clickButton() {
            var button = document.getElementById("test");
            var buttonX = button.offsetLeft + button.offsetWidth / 2;
            var buttonY = button.offsetTop + button.offsetHeight / 2;
            if (window.eventSender) {
                eventSender.mouseMoveTo(buttonX, buttonY);
                eventSender.mouseDown();
                eventSender.mouseUp();
            }
        }
    </script>
</head>
<body onload="clickButton()">
<button id="test" onclick="clickHandler()"></button>
<div id="console"></div>
</body>
</html>
